feat: 멍스타그램 상세 모달 반응형 (#131)
