+++ src/app/(providers)/@modal/_components/index.ts
@@ -1,0 +1,5 @@
+import CommentForm from './CommentForm';
+import CommentList from './CommentList';
+import LikeButton from './LikeButton';
+
+export { CommentForm, CommentList, LikeButton };
