+++ src/app/globals.css
@@ -171,6 +171,7 @@ a {
 .slick-dots li button:before {
   font-size: 10px !important;
 }
-.slick-prev:before, .slick-next:before {
+.slick-prev:before,
+.slick-next:before {
   color: #eee !important;
 }
